+++ src/main/java/fr/marcwrobel/jbanking/bic/Bic.java
@@ -104,7 +104,7 @@ public final class Bic {
   private static boolean hasKnownCountryCode(String s) {
     return IsoCountry.fromAlpha2Code(
             s.substring(COUNTRY_CODE_INDEX, COUNTRY_CODE_INDEX + COUNTRY_CODE_LENGTH))
-        != null;
+        .isPresent();
   }
 
   /**